refactor: IndexContentReader 패키지 옮긴다
