+++ src/test/java/org/gsh/genidxpage/service/IndexPageGeneratorTest.java
@@ -4,6 +4,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
 import org.assertj.core.api.Assertions;
+import org.gsh.genidxpage.repository.IndexContentReader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
